Allow image/bmp files to be uploaded as image/png
The .bmp file selected or drag-and-dropped is converted to image/png format
by the front-end side before upload, so that it does not affect the back-end
side.

It seems user-friendly for some Touhou game players who want to upload their
gameplay snapshots. Because the snapshot produced by pressing the "p" or "Home"
key during gameplay is saved as a .bmp file.